+++ 2019/5/Part2.cs
@@ -6,7 +6,7 @@ namespace aoc.y2019.day5
     {
         public Part2(string file, int exp) : base(file, exp) { }
 
-        protected override int doWork() {
+        protected override long doWork() {
             return base.doWork(new Provider(5));
         }
     }
